src/main/kotlin/com/dls/insuranceadvisorservice/domain/insurancelines/rules/MinimumIncomeRule.kt
package com.dls.insuranceadvisorservice.domain.insurancelines.rules

import com.dls.insuranceadvisorservice.domain.insurancelines.RiskScoreRule
import com.dls.insuranceadvisorservice.domain.RiskProfileForInsuranceLine
import com.dls.insuranceadvisorservice.domain.UserRiskProfile

class MinimumIncomeRule: RiskScoreRule {

     override fun execute(userRiskProfile: UserRiskProfile, riskProfileForInsuranceLine: RiskProfileForInsuranceLine){
        if(userRiskProfile.income <= 0){
            riskProfileForInsuranceLine.makeIneligible()

        }
    }
}